pyproject.toml
[tool.poetry]
name = "transactions-storage"
version = "0.1.0"
description = ""
authors = ["aleksandr.rybin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.9"
sqlalchemy = "^2.0.38"
pydantic = "^2.10.6"
asyncpg = "^0.30.0"
aiosqlite = "^0.21.0"
uvicorn = "^0.34.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"
asyncmock = "^0.4.2"
pytest-cov = "^6.0.0"
alembic = "^1.14.1"
pydantic-settings = "^2.8.1"
celery = "^5.4.0"
psycopg2 = "^2.9.10"
pytest = "^8.3.5"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
